+++ package.json
@@ -6,8 +6,8 @@
   "type": "module",
   "scripts": {
     "build": "webpack",
-    "start:dev": "nodemon --loader ts-node/esm ./src/index.ts",
-    "start:prod": "npm run build && node dist/build.js",
+    "start:dev": "node --experimental-specifier-resolution=node --loader ts-node/esm ./src/index.ts",
+    "start:prod": "npm run build && node dist/build.cjs",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
